Ecohub/requirements.txt
Django==4.2
djangorestframework
django-allauth
Pillow
stripe
django-crispy-forms
crispy-bootstrap5
psycopg2-binary
python-decouple
django-filter
django-extensions
gunicorn
whitenoise

djangorestframework-simplejwt
mysqlclient
python-decouple
django-allauth
django-crispy-forms
crispy-bootstrap5
django-cors-headers
django-filter